+++ mars/dataframe/reduction/tests/test_reduction_execute.py
@@ -657,129 +657,5 @@ class TestAggregate(TestBase):
                                        data.agg(all_aggs))
 
 
-class TestCorr(TestBase):
-    def setUp(self):
-        self.executor = ExecutorForTest()
-
-    def testDataFrameCorr(self):
-        rs = np.random.RandomState(0)
-        raw = rs.rand(20, 10)
-        raw = pd.DataFrame(np.where(raw > 0.4, raw, np.nan), columns=list('ABCDEFGHIJ'))
-        raw['k'] = pd.Series(['aaa'] * 20)
-
-        df = from_pandas_df(raw)
-
-        result = df.corr()
-        pd.testing.assert_frame_equal(self.executor.execute_dataframe(result, concat=True)[0],
-                                      raw.corr())
-
-        result = df.corr(method='kendall')
-        pd.testing.assert_frame_equal(self.executor.execute_dataframe(result, concat=True)[0],
-                                      raw.corr(method='kendall'))
-
-        df = from_pandas_df(raw, chunk_size=6)
-
-        with self.assertRaises(Exception):
-            self.executor.execute_dataframe(df.corr(method='kendall'), concat=True)
-
-        result = df.corr()
-        pd.testing.assert_frame_equal(self.executor.execute_dataframe(result, concat=True)[0],
-                                      raw.corr())
-
-        result = df.corr(min_periods=7)
-        pd.testing.assert_frame_equal(self.executor.execute_dataframe(result, concat=True)[0],
-                                      raw.corr(min_periods=7))
-
-    def testDataFrameCorrWith(self):
-        rs = np.random.RandomState(0)
-        raw_df = rs.rand(20, 10)
-        raw_df = pd.DataFrame(np.where(raw_df > 0.4, raw_df, np.nan), columns=list('ABCDEFGHIJ'))
-        raw_df2 = rs.rand(20, 10)
-        raw_df2 = pd.DataFrame(np.where(raw_df2 > 0.4, raw_df2, np.nan), columns=list('ACDEGHIJKL'))
-        raw_s = rs.rand(20)
-        raw_s = pd.Series(np.where(raw_s > 0.4, raw_s, np.nan))
-        raw_s2 = rs.rand(10)
-        raw_s2 = pd.Series(np.where(raw_s2 > 0.4, raw_s2, np.nan), index=raw_df2.columns)
-
-        df = from_pandas_df(raw_df)
-        df2 = from_pandas_df(raw_df2)
-
-        result = df.corrwith(df2)
-        pd.testing.assert_series_equal(self.executor.execute_dataframe(result, concat=True)[0],
-                                       raw_df.corrwith(raw_df2))
-
-        result = df.corrwith(df2, axis=1)
-        pd.testing.assert_series_equal(self.executor.execute_dataframe(result, concat=True)[0],
-                                       raw_df.corrwith(raw_df2, axis=1))
-
-        result = df.corrwith(df2, method='kendall')
-        pd.testing.assert_series_equal(self.executor.execute_dataframe(result, concat=True)[0],
-                                       raw_df.corrwith(raw_df2, method='kendall'))
-
-        df = from_pandas_df(raw_df, chunk_size=4)
-        df2 = from_pandas_df(raw_df2, chunk_size=6)
-        s = from_pandas_series(raw_s, chunk_size=5)
-        s2 = from_pandas_series(raw_s2, chunk_size=5)
-
-        with self.assertRaises(Exception):
-            self.executor.execute_dataframe(df.corrwith(df2, method='kendall'), concat=True)
-
-        result = df.corrwith(df2)
-        pd.testing.assert_series_equal(self.executor.execute_dataframe(result, concat=True)[0].sort_index(),
-                                       raw_df.corrwith(raw_df2).sort_index())
-
-        result = df.corrwith(df2, axis=1)
-        pd.testing.assert_series_equal(self.executor.execute_dataframe(result, concat=True)[0].sort_index(),
-                                       raw_df.corrwith(raw_df2, axis=1).sort_index())
-
-        result = df.corrwith(s)
-        pd.testing.assert_series_equal(self.executor.execute_dataframe(result, concat=True)[0].sort_index(),
-                                       raw_df.corrwith(raw_s).sort_index())
-
-        result = df.corrwith(s2, axis=1)
-        pd.testing.assert_series_equal(self.executor.execute_dataframe(result, concat=True)[0].sort_index(),
-                                       raw_df.corrwith(raw_s2, axis=1).sort_index())
-
-    def testSeriesCorr(self):
-        rs = np.random.RandomState(0)
-        raw = rs.rand(20)
-        raw = pd.Series(np.where(raw > 0.4, raw, np.nan))
-        raw2 = rs.rand(20)
-        raw2 = pd.Series(np.where(raw2 > 0.4, raw2, np.nan))
-
-        s = from_pandas_series(raw)
-        s2 = from_pandas_series(raw2)
-
-        result = s.corr(s2)
-        self.assertEqual(self.executor.execute_dataframe(result, concat=True)[0],
-                         raw.corr(raw2))
-
-        result = s.corr(s2, method='kendall')
-        self.assertEqual(self.executor.execute_dataframe(result, concat=True)[0],
-                         raw.corr(raw2, method='kendall'))
-
-        result = s.autocorr(2)
-        self.assertEqual(self.executor.execute_dataframe(result, concat=True)[0],
-                         raw.autocorr(2))
-
-        s = from_pandas_series(raw, chunk_size=6)
-        s2 = from_pandas_series(raw2, chunk_size=4)
-
-        with self.assertRaises(Exception):
-            self.executor.execute_dataframe(s.corr(s2, method='kendall'), concat=True)
-
-        result = s.corr(s2)
-        self.assertAlmostEqual(self.executor.execute_dataframe(result, concat=True)[0],
-                               raw.corr(raw2))
-
-        result = s.corr(s2, min_periods=7)
-        self.assertAlmostEqual(self.executor.execute_dataframe(result, concat=True)[0],
-                               raw.corr(raw2, min_periods=7))
-
-        result = s.autocorr(2)
-        self.assertAlmostEqual(self.executor.execute_dataframe(result, concat=True)[0],
-                               raw.autocorr(2))
-
-
 if __name__ == '__main__':  # pragma: no cover
     unittest.main()
